Fix dependency cycle with forward declarations
