+++ source/Cell.h
@@ -4,23 +4,7 @@
 
 #include "Config.h"
 #include "Sequence.h"
-
-struct Trait{
-
-  public:
-
-    Trait()
-    : guess(-1)
-    , fitness(0)
-    , sense_idx(0)
-    { ;}
-
-    int guess;
-    double fitness;
-    emp::Ptr<Sequence> seq;
-    size_t sense_idx;
-
-};
+#include "Trait.h"
 
 struct Cell{
