+++ source/Trait.h
@@ -1,0 +1,21 @@
+#pragma once
+
+// forward declaration
+class Sequence;
+
+struct Trait{
+
+  public:
+
+    Trait()
+    : guess(-1)
+    , fitness(0)
+    , sense_idx(0)
+    { ;}
+
+    int guess;
+    double fitness;
+    emp::Ptr<Sequence> seq;
+    size_t sense_idx;
+
+};
